// src/types/description.ts
export type DescrParams = {
    name_product: string
    target_audience?: string;
    product_category_analysis?: string;
    consumer_objections?: string;
    consumer_expectations?: string
    purchase_triggers?: string;
    product_keywords?: string;
    product_characteristics?: string;
    user_prompt?:string;
}